Update i18n module to better handle dicts. Expose dicts on Lock instance for easy override
