+++ lib/i18n/index.js
@@ -1,63 +1,37 @@
-var dics_data = {
-  'ar': require('../../i18n/ar.json'),
-  'da': require('../../i18n/da.json'),
-  'de': require('../../i18n/de.json'),
-  'en': require('../../i18n/en.json'),
-  'es': require('../../i18n/es.json'),
-  'fr': require('../../i18n/fr-FR.json'),
-  'fr-FR': require('../../i18n/fr-FR.json'),
-  'he': require('../../i18n/he.json'),
-  'it': require('../../i18n/it.json'),
-  'ja': require('../../i18n/ja.json'),
-  'ko': require('../../i18n/ko.json'),
-  'nb-NO': require('../../i18n/nb-NO.json'),
-  'nl': require('../../i18n/nl-NL.json'),
-  'nl-NL': require('../../i18n/nl-NL.json'),
-  'pt': require('../../i18n/pt.json'),
-  'pt-BR': require('../../i18n/pt-BR.json'),
-  'ru': require('../../i18n/ru.json'),
-  'sv': require('../../i18n/sv.json'),
-  'tlh': require('../../i18n/tlh.json'),
-  'tr': require('../../i18n/tr.json'),
-  'zh': require('../../i18n/zh.json'),
-  'zh-TW': require('../../i18n/zh-TW.json')
-};
+/**
+ * Module dependencies.
+ */
 
-var default_dict = dics_data['en'];
+var Dictionary = require('./dictionary');
+var dicts = require('./dicts');
+var default_dict = dicts['en'];
+var _ = require('underscore');
 
-function findProp(o, s) {
-    s = s.replace(/\[(\w+)\]/g, ':$1'); // convert indexes to properties
-    s = s.replace(/^\:/, '');           // strip a leading dot
-    var a = s.split(':');
-    while (a.length) {
-        var n = a.shift();
-        if (n in o) {
-            o = o[n];
-        } else {
-            return;
-        }
-    }
-    return o;
-}
+/**
+ * Expose all supported `dicts`
+ */
 
-function Dictionary (data) {
-  this._data = data;
-}
+module.exports.dicts = dicts;
 
-Dictionary.prototype.t = function (key) {
-  return findProp(this._data, key) || findProp(default_dict, key);
-};
+/**
+ * Retrieve a new `Dictionary` from `lang`
+ *
+ * @param {String|Object} lang
+ * @return {Dictionary}
+ * @public
+ */
 
-module.exports.getDict = function ( langOrDict ) {
+module.exports.getDict = function (lang) {
 
-  if (!langOrDict) {
+  if (!lang) {
     return new Dictionary(default_dict);
   }
 
-  if ( typeof langOrDict === 'string' ) {
-    var dict = dics_data[langOrDict] || dics_data[langOrDict.split('-')[0]];
+  if ('string' === typeof lang) {
+    var dict = dicts[lang] || dicts[lang.split('-')[0]];
     return new Dictionary(dict);
-  } else {
-    return new Dictionary(langOrDict);
   }
-};
+
+  return new Dictionary(lang);
+
+}
